Moved ProductImageUrlMap to the `util` package
